intermediate PR, PC tweakinnn
