+++ otr/Assets/CarController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7bda998fc30c8440f95bc045f06714f2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
