+++ otr/Assets/CopFollowDebug.cs
@@ -1,0 +1,26 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class CopFollowDebug : MonoBehaviour
+{
+    // Start is called before the first frame update
+    public UnityEngine.AI.NavMeshAgent agentToDebug;
+    private LineRenderer lineRenderer;
+    void Start()
+    {
+        lineRenderer = GetComponent<LineRenderer>();
+    }
+
+    // Update is called once per frame
+    void Update()
+    {
+        if (agentToDebug.hasPath) {
+            lineRenderer.positionCount = agentToDebug.path.corners.Length;
+            lineRenderer.SetPositions(agentToDebug.path.corners);
+            lineRenderer.enabled = true;
+        } else {
+            lineRenderer.enabled = false;
+        }
+    }
+}
