+++ otr/Assets/CopFollowDebug.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 38dc5a06336e44fb4b102feacce0654d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
