+++ otr/Assets/Scripts/Cars/PlayerCarController 2.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fb50a65144be04fccb4bc492b777b7c3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
