--- otr/Assets/Scripts/Cars/PlayerCarController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e95c8c319eb53465eb293a3c469d444b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
