+++ otr/Assets/Scripts/CopFollow.cs
@@ -6,8 +6,10 @@ using UnityEngine.AI;
 public class CopFollow : MonoBehaviour
 {
     private NavMeshAgent Cop;
+    public GameObject respawnableCop;
     public GameObject Player;
     public float MobDistanceRun = 4.0f;
+
 
     // Start is called before the first frame update
     void Start()
