+++ otr/Assets/nav.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1dd71a239a152496b82ba27c17c3bee4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
